requirements.txt
numpy
opencv-python
json-tricks
torch
xtcocotools
mmcv-full>=1.5.3
